Reorganize anyhow impls
