+++ crates/cgp-error-anyhow/src/impls/use_anyhow_error.rs
@@ -1,0 +1,8 @@
+use anyhow::Error;
+use cgp_core::error::ProvideErrorType;
+
+pub struct UseAnyhowError;
+
+impl<Context> ProvideErrorType<Context> for UseAnyhowError {
+    type Error = Error;
+}
